setup.py: Rename the binary to gdbus-codegen-glibmm2
In order to allow coinstallation with the 1.x versions, we need to
rename the installed script file and ensure that the installed python
module gets installed in its own directory. We achieve the latter by
changing the version number, since this affects the location where the
module files are installed.