templates:: Call error initialization method in proxy constructor
This makes it so it is not needed to call it manually before using a proxy
to get Error exceptions to be thrown instead of plain Glib::Error
exceptions.

In stubs this is not needed since Error instances are created when
returning errors in a method invocation and Error calls
register_domain_with_dbus() in its constructor.