templates: Add a changed signal for each property
That can be used to find out when a D-Bus property changes value.

Do not include value as argument since the <property name>_get() method can
be used to get it (and value is cached by Glib).

Wrap sigc::signal in a getter method so it is possible to change to
creating signals when they are first connected to in the future without
breaking API.

Looking at the many-types test I wonder if it would be worth to have a map
for looking up the sigc::signal from a property name. But having that many
properties is probably not that common... ?