templates: Add timeout argument to proxy D-Bus method calls
Needed because some services do not return until a potentially very long
operation has finished (e.g. waiting for input from user or configuring
some hardware on the system).

Default to -1 which means use the GDBusProxy default (there is a property
that is also set to -1 by default and means "use hard coded value of 25s
in GDBusConnection").