templates: Add sync versions of proxy method calls
Value is just returned if method has a single output argument.

If a method has more than a single output argument the values are packed in a
std::tuple. A bit cumbersome to use prior to C++17 perhaps, but C++17 makes
it nicer with structured bindings.

The many_types.xml test has a method with 16 out parameters. The std::tuple
takes up many columns in that case, but decided to ignore it since it is
a bit of an extreme case.