setup.py: Rename the binary to gdbus-codegen-glibmm3
In order to allow coinstallation with the older versions, we need to
rename the installed script file and ensure that the installed python
module gets installed in its own directory. We achieve the latter by
changing the version number, since this affects the location where the
module files are installed.