tests: Fix Variant2 test in stub of integration test
Without this the invocation.ret() call in TestImpl::TestVariant2()
will print:

(process:9052): GLib-GIO-WARNING **: 12:38:51.251: Type of return value is incorrect: expected '(sv)', got '(ss)''

That is, two strings are returned while TestVariant2 has a string and a
variant as out parameters.

The creation of variantValue is copied from TestImpl::TestVariant() above.

on_test_variant2_finished() was never called on the client side before the
tests finished. (No timeout strikes on client side since it is 25s by
default and tests are finished before that.) Now the callback is called on
the client side but it fails. Will be fixed in following commit.