Add Proxy suffix to proxies and Stub suffix to stubs
Avoids name clashes between namespace names and class names when an
interface is a put "below" another one. E.g. the interfaces a.B and a.B.C
would add a class named B for a.B and a namespace called B for a.B.C.

Since the proxy and stub classes are named differently now it is also OK
to include *_proxy.h and *_stub.h in the same compilation unit. The use
case for this would be e.g. a service with name a.B requiring a client to
implement an interface with the name a.B.C .

Fixes #57 .