Make stub follow rule of five
And delete copy/move constructors/assignment operators in stub.

The stub connects to signals with sigc::mem_fun. If the object should be
movea:ble/copy:able this needs to be taken care of.

Stub should probably never be copy:able but it might make sense to make it
move:able. Have not had a need for that yet though so just delete move
ctor and assignment operator as well.