Make stub inherit from sigc::trackable
Since it connects signals to sigc::mem_fun slots. Inheriting from
sigc::trackable will make it so signals are automatically disconnected
when object is destroyed.

Inheriting from sigc::trackable does _not_ make move operations work as
wanted. sigc::trackable just disconnects the object that is moved from. It
does not move over the signal connections to the moved to object. Still
needs to be taken care of manually if object should be move:able. No need
for that at the moment though.