Use object id to check if object is already registered in stub
Long term plan is to make it possible to use the same stub instance for
multiple object paths. See #81 . At the moment there is only a single id
stored in a stub instance though and it will "leak" if we allow
overwriting it, so this is currently the best/simplest way to prevent
that "leak".