proxy: Add possibility to check result in <property>_get()
Add the possibility to pass a bool pointer to the methods used to get
property values. This will be set to true if the property could be read,
and otherwise set to false.

Properties can be optional, so even if they appear in the interface XML
they might not be present in the object the proxy is created for. By
adding the possibility to check if the property could be read, the
caller now have a way of telling if the returned value is valid or not.

This fixes #83.