proxy: Remove string lookup when getting properties
There is no need to first get a list of string for the properties that
are in the property cache, then try to find the property of interest in
that list before doing get_cached_property(). If the property is not in
the cache, get_cached_property will return an invalid Variant.

This fixes #84.