stub: Allow registering the same instance on multiple connections/paths
Stubs can be registered on multiple connections, and also multiple times
on the same connection, as long as object paths are different. This
commit enables this, by turning the object registration data into an
array.

Note that signals need to be emitted on each connection/path.

A check to avoid duplicate registrations (with the same connection/path)
is not needed, as it's already performed by
g_dbus_connection_register_object().

Fixes: #81

Signed-off-by: Alberto Mardegan <amardegan@luxoft.com>